+++ src/App.js
@@ -1,4 +1,4 @@
-import UserForm from "./components/UserForm.js";
+import UserForm from "./components/form/UserForm";
 
 export default function App() {
